requirements.txt
google-genai
gitpython
psutil
python-dotenv
numpy
SpeechRecognition
openai-whisper
pyaudio
pyserial